Use new model for Apple receipt query view.
